redesigned caching of user data, now the cache is bound to user login instead of token, which eliminates errors of data irrelevance.

+++ mindfulguard/safe/delete.py
@@ -15,17 +15,17 @@ class Delete(SafeBase):
             self._model_token.token = token
             self._model_safe.id = safe_id
             db = self._pgsql_safe.delete(self._model_token, self._model_safe)
+            db_user_info = self._pgsql_user.get_information(self._model_token)
             await self._connection.open()
             await db.execute()
+            await db_user_info.execute()
             self._status_code = db.status_code
-            if db.status_code == OK:
-                db_user_info = self._pgsql_user.get_information(self._model_token)
-                await db_user_info.execute()
+            if db.status_code == OK and db_user_info.status_code == OK:
                 self._s3.set_bucket_name(db_user_info.response.login)
                 object_list = [obj.object_name for obj in self._s3.object().get_all_objects(f'{self._s3.CONTENT_PATH}/{safe_id}/')]
                 self._s3.object().delete_objects(object_list) # type: ignore
                 for i in self._redis.client().connection.scan_iter(
-                    f'{self._model_token.token}:{self._redis.PATH_SAFE_ALL_ITEM}'
+                    f'{db_user_info.response.login}:{self._redis.PATH_SAFE_ALL_ITEM}'
                 ):
                     self._redis.client().connection.delete(i)
             return
